build mail message view toolbar
